Exception handler

+++ src/main/java/airport/controller/TouristController.java
@@ -1,10 +1,8 @@
 package airport.controller;
 
 import airport.exception.TouristNotFoundException;
-import airport.model.Flight;
 import airport.model.Lot;
 import airport.model.Tourist;
-import airport.service.FlightService;
 import airport.service.LotService;
 import airport.service.TouristService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,22 +10,24 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.ArrayList;
+
+import javax.validation.Valid;
 import java.util.List;
+import java.util.Optional;
 
 @RequestMapping(value = "/tourist")
 @RestController
+@CrossOrigin
 public class TouristController {
 
     private final TouristService touristService;
     private final LotService lotService;
-    private final FlightService flightService;
+
 
     @Autowired
-    public TouristController(TouristService touristService, LotService lotService, FlightService flightService) {
+    public TouristController(TouristService touristService, LotService lotService) {
         this.touristService = touristService;
         this.lotService = lotService;
-        this.flightService = flightService;
     }
 
     @GetMapping(value = "/all")
@@ -35,66 +35,52 @@ public class TouristController {
         return touristService.getAllTourists();
     }
 
-    @GetMapping(value = "/{name}")
+    @GetMapping(value = "/name/{name}")
     public List<Tourist> getName(@PathVariable("name") String name) {
         return touristService.getByName(name);
     }
 
-    @GetMapping(value = "/id/{id}")
-    public Tourist showone(@PathVariable("id") Integer id) {
-        return touristService.findyById(id).orElseThrow(() -> new TouristNotFoundException(id));
+    @GetMapping(value = "/{id}")
+    public Tourist getOne(@PathVariable("id") Integer id) {
+        return touristService.getOneById(id).orElseThrow(() -> new TouristNotFoundException(id));
+    }
+
+    @PostMapping(value = "/add")
+    public Tourist addTourist(@RequestBody @Valid Tourist tourist) {
+        return touristService.save(tourist);
+    }
+
+    @PostMapping(value = "/lot/{id}")
+    public ResponseEntity<Lot> addTouristToFlight(@RequestBody Tourist tourist, @PathVariable Integer id) {
+        Lot lot1 = null;
+        Optional<Lot> lot = lotService.getOneByID(id);
+
+        if (lot.isPresent()) {
+            Integer flightId = lot.get().getFlight().getFlightId();
+            lot1 = lot.get();
+            tourist.setFlightId(flightId);
+            List<Tourist> idTourist = lot1.getTouristList();
+
+            if (idTourist.size() < 16)
+                idTourist.add(tourist);
+            else
+                return new ResponseEntity<Lot>(HttpStatus.NOT_ACCEPTABLE);
+
+            Lot lots = lotService.addLot(lot1);
+            return new ResponseEntity<Lot>(lots, HttpStatus.CREATED);
+        } else {
+            return new ResponseEntity<Lot>(HttpStatus.NOT_FOUND);
+        }
     }
 
     @DeleteMapping(value = "/{id}")
     public ResponseEntity<String> deleteTouristById(@PathVariable Integer id) {
-        touristService.deleteById(id);
-        return new ResponseEntity<String>(HttpStatus.OK);
-    }
-
-    @PostMapping(value = "/add")
-    public Tourist addTourist(@RequestBody Tourist tourist) {
-        return touristService.save(tourist);
-    }
-
-
-//    @GetMapping(value = "/wow")
-//    public Optional<Tourist> afaf() {
-//        Lot lot1 = lotService.showOne(2);
-//        List<Tourist> idTourist = lot1.getTouristList();
-//        Integer lng = idTourist.size();
-//        Optional<Tourist> tourist = touristService.findyById(4);
-//        tourist.get().setFlightId(lng);
-//        return tourist;
-//    }
-
-
-    @PostMapping(value = "/add/{id}")
-    public ResponseEntity<Lot> addTouristToFlight(@RequestBody Tourist tourist, @PathVariable Integer id) {
-        Lot lot1 = lotService.showOne(id);
-        tourist.setFlightId(id);
-        List<Tourist> idTourist = lot1.getTouristList();
-
-        if(idTourist.size() < 16)
-            idTourist.add(tourist);
-        else
-            return new ResponseEntity<Lot>(HttpStatus.NOT_ACCEPTABLE);
-
-        Lot lots = lotService.addLot(lot1);
-        return  new ResponseEntity<Lot>(lots,HttpStatus.CREATED);
-    }
-
-    @PostMapping(value = "/zero/{id}")
-    public ResponseEntity<Lot> addToEmptyFlight(@RequestBody Tourist tourist, @PathVariable Integer id) {
-        List<Tourist> touristList = new ArrayList<>();
-        tourist.setFlightId(id);
-        touristList.add(tourist);
-
-        Flight flightById = flightService.getFlightById(id);
-        Lot newLot = new Lot(flightById,touristList);
-
-        Lot lott = lotService.addLot(newLot);
-
-        return  new ResponseEntity<Lot>(lott,HttpStatus.CREATED);
+        if(touristService.getOneById(id).isPresent()) {
+            touristService.deleteById(id);
+            return new ResponseEntity<String>(HttpStatus.OK);
+        } else {
+            return new ResponseEntity<String>(HttpStatus.NOT_FOUND);
+        }
     }
 }
